refactor: move type system to core/types/
- Copy xvr_type.h and xvr_type.c to src/core/types/
- Update includes to use relative paths within core/
- Create redirect files in src/ that include the core versions
- Update CMakeLists.txt to compile from core/types/
- Fix header guard collision (XVR_TYPE_H -> XVR_CORE_TYPE_H)
